search sends to sentence bar, fix loading issue on created words, delete button on sentence bar fixed to the right correctly, move screen folders for better organization
