+++ package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "@react-native-community/masked-view": "^0.1.10",
     "expo": "~38.0.8",
+    "expo-av": "^8.3.0",
     "expo-constants": "^9.1.1",
     "expo-file-system": "~9.0.1",
     "expo-image-picker": "^8.3.0",
